Services refactored and added actions section to tables
