+++ src/services/DbService.ts
@@ -1,0 +1,10 @@
+import { Client } from "pg"
+
+class DbService {
+  client: Client
+  constructor (client: Client) {
+    this.client = client
+  }
+}
+
+export default DbService
