+++ src/services/TimesService.ts
@@ -1,35 +1,6 @@
-import { Client } from 'pg'
+import DbService from "./DbService";
 
-class DbService {
-  client: Client
-  constructor (url: string) {
-    this.client = new Client({
-      connectionString: url,
-      ssl: process.env.DATABASE_SSL ? {
-        rejectUnauthorized: false
-      } : false
-    })
-    this.client.connect().catch((err) => {
-      console.error(err)
-    })
-  }
-
-  async getOwners () {
-    return await this.client.query('select * from owner order by id desc')
-  }
-
-  async getUsers () {
-    return await this.client.query('select * from "user"')
-  }
-
-  async getProjects (ownerId: string) {
-    return await this.client.query(`select * from project where owner_id = ${ownerId}`)
-  }
-
-  async createOwner (name: string) {
-    return await this.client.query(`insert into "owner"(name, active) values ('${name}', true)`)
-  }
-
+class TimeService extends DbService{
   async getTimes (
     name = '',
     owner = '',
@@ -85,49 +56,6 @@ class DbService {
       and "end" is null`)
   }
 
-  async getProjectsDetail () {
-    return await this.client.query(`
-    select
-      p.id,
-      u.id as time_id,
-      p.name, 
-      p.description,
-      u.name as "user",
-      user_hours as "hours",
-      project_hours,
-      ((user_hours + 1) / (p.budget + 0.1)) * 100 as "percent"
-    from project p
-    inner join owner o on o.id = p.owner_id
-    inner join (select 
-        u.id as id,
-        u.name as name,
-        t.project_id,
-        sum(ROUND(cast((extract(epoch from "end" - "start")) as numeric) / 3600, 2)) as user_hours
-      from time t
-      inner join "user" u on u.id = user_id
-      group by u.id, u.name, t.project_id) u on u.project_id = p.id
-    inner join (
-      select 
-        p.id as project_id, 
-        sum(ROUND(cast((extract(epoch from "end" - "start")) as numeric) / 3600, 2)) as project_hours
-      from project p
-      inner join time t on t.project_id = p.id
-      group by p.id) as ih on ih.project_id = p.id
-    `)
-  }
-
-  async createProject (ownerId: number, name: string, description: string, budget: number) {
-    return await this.client.query(`
-      insert into public.project(
-      owner_id, name, description, budget)
-      values (${ownerId}, '${name}', '${description}', ${budget})
-    `)
-  }
-
-  async createUser (name: string, username: string, password: string) {
-    return await this.client.query(`insert into "user"(name, username, password, active) values ('${name}', '${username}', '${password}', true)`)
-  }
-
   async getTodayUser (userId: number) {
     return await this.client.query(`
       select 
@@ -171,4 +99,4 @@ class DbService {
   }
 }
 
-export default DbService
+export default TimeService
